+++ Home/Models/Activity.swift
@@ -1,0 +1,20 @@
+//
+//  Activity.swift
+//  FitnessAppMero
+//
+//  Created by Meriç Cem Baysar on 22.07.2024.
+//
+
+import SwiftUI
+
+
+
+struct Activity {
+    let id : Int
+    let title: String
+    let subtitle: String
+    let image: String
+    let tintColor: Color
+    let amount: String
+    
+}
