+++ Home/Models/Workout.swift
@@ -1,0 +1,20 @@
+//
+//  Workout.swift
+//  FitnessAppMero
+//
+//  Created by Meriç Cem Baysar on 22.07.2024.
+//
+
+import SwiftUI
+
+
+struct Workout {
+    let id: Int
+    let title: String
+    let image: String
+    let tintColor: Color
+    let duration: String
+    let date: String
+    let calories: String
+
+}
